+++ src/Attributes/Response204NoContentAttribute.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+
+namespace PowerUtils.AspNetCore.Extensions.Attributes
+{
+    public class Response204NoContentAttribute : ProducesResponseTypeAttribute
+    {
+        public Response204NoContentAttribute() : base(StatusCodes.Status204NoContent) { }
+    }
+}