+++ src/Attributes/Response206PartialContentAttribute.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+
+namespace PowerUtils.AspNetCore.Extensions.Attributes
+{
+    public class Response206PartialContentAttribute : ProducesResponseTypeAttribute
+    {
+        public Response206PartialContentAttribute(Type type) : base(type, StatusCodes.Status206PartialContent) { }
+    }
+}