+++ src/Attributes/Response208AlreadyReportedAttribute.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+
+namespace PowerUtils.AspNetCore.Extensions.Attributes
+{
+    public class Response208AlreadyReportedAttribute : ProducesResponseTypeAttribute
+    {
+        public Response208AlreadyReportedAttribute(Type type) : base(type, StatusCodes.Status208AlreadyReported) { }
+    }
+}