+++ tests/PowerUtils.AspNetCore.Attributes.ResponseTypes.Tests/FakeResponse.cs
@@ -1,5 +1,0 @@
-﻿namespace PowerUtils.AspNetCore.Attributes.ResponseTypes.Tests;
-public class FakeResponse
-{
-    public int Val { get; set; }
-}