+++ tests/PowerUtils.AspNetCore.Extensions.Tests/Fakes/FakeResponse.cs
@@ -1,0 +1,5 @@
+﻿namespace PowerUtils.AspNetCore.Extensions.Tests.Fakes;
+public class FakeResponse
+{
+    public int Val { get; set; }
+}